refactor: extract global helpers to src/app/ module
- Create src/app/helpers.rs with shared helper functions
- Move apply_to_all_views, sync_visual_prefs_to_all_views, lang_is_es
- Add redraw_all_views for future use
- Remove duplicate functions from main.rs (-58 lines)
- This unblocks extraction of input and ui modules
